move frmUserinformation to views
