+++ OCMS/OCMS.csproj
@@ -100,6 +100,12 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="VIEW\frmUsersInformation.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="VIEW\frmUsersInformation.Designer.cs">
+      <DependentUpon>frmUsersInformation.cs</DependentUpon>
+    </Compile>
     <EmbeddedResource Include="frmLogin.resx">
       <DependentUpon>frmLogin.cs</DependentUpon>
     </EmbeddedResource>
@@ -124,6 +130,9 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <EmbeddedResource Include="VIEW\frmUsersInformation.resx">
+      <DependentUpon>frmUsersInformation.cs</DependentUpon>
+    </EmbeddedResource>
     <None Include="OCMSDataSet.xsc">
       <DependentUpon>OCMSDataSet.xsd</DependentUpon>
     </None>
